requirements.txt
fastapi==0.79.0
pydantic[email]==1.9.1
uvicorn==0.18.2
sqlalchemy==1.4.39
psycopg2-binary==2.9.3
passlib[bcrypt]==1.7.4
python-dotenv==0.20.0
pyjwt==2.4.0
cryptography==37.0.4
